+++ isis/src/lro/apps/lronac2pds/lronac2pds.cpp
@@ -241,7 +241,7 @@ void OutputLabel ( std::ofstream &fout, Cube* cube ) {
     stream << labelPvl;
     pdsLabel += stream.str().c_str();
 
-    while (pdsLabel.length() < (unsigned int) (labelRecords * recordBytes)) {
+    while ((int)pdsLabel.length() < (int) (labelRecords * recordBytes)) {
         pdsLabel += '\n';
     }
